% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runShinyApp}
\alias{runShinyApp}
\title{Launch Shiny app to visualize Healthy Eating Index (HEI) scores}
\usage{
runShinyApp()
}
\description{
This function launches a Shiny application that allows users to visualize HEI scores calculated from National Health and Nutrition Examination Survey (NHANES) 24-hour dietary recall data.
}
\examples{
library(tidyr)
library(dplyr)
library(stringr)
library(shiny)
library(shinythemes)
library(ggplot2)
library(tibble)
library(ggpubr)
library(grDevices)
library(RColorBrewer)
library(rlang)

runShinyApp()

}
